Move example code
