+++ traitsui/qt4/extra/styled_tree_editor.py
@@ -127,66 +127,3 @@ class StyledTreeNode(TreeNode):
     """
     #: Set to True to use ItemDelegate for drawing
     delegates_style = Bool(False)
-
-
-if __name__ == '__main__':
-    from random import choice
-    from traits.api import HasTraits, Instance, List, Unicode
-    from traitsui.api import UItem, View
-
-    class MyDataItem(HasTraits):
-        text = Unicode
-
-    class MyData(HasTraits):
-        name = Unicode('Rooty McRootface')
-        items = List(Instance(MyDataItem))
-
-        def _items_default(self):
-            DATA_ITEMS = (
-                'I live on\nmultiple\nlines!', 'Foo\nBar', 'Baz', 'Qux',
-            )
-            return [MyDataItem(text=choice(DATA_ITEMS)) for _ in range(5)]
-
-    class MyStyledTreeItemDelegate(StyledTreeItemDelegate):
-        def custom_paint(self, painter, option, index, node, instance):
-            """ The paint() method for subclasses to override.
-            """
-            text = node.get_label(instance)
-            rect = painter.drawText(option.rect.left(),
-                                    option.rect.top() + 2,
-                                    option.rect.width(),
-                                    option.rect.height() + 2,
-                                    QtCore.Qt.TextWordWrap, text)
-
-            # Need to set the appropriate sizeHint of the item.
-            size = QtCore.QSize(rect.width(), rect.height() + 4)
-            self.updateSizeHint(index, size)
-
-    class MyClass(HasTraits):
-        root = Instance(MyData, args=())
-        traits_view = View(
-            UItem(
-                'root',
-                editor=StyledTreeEditor(
-                    item_delegate_klass=MyStyledTreeItemDelegate,
-                    nodes=[
-                        StyledTreeNode(
-                            node_for=[MyData],
-                            children='items',
-                            label='name',
-                        ),
-                        StyledTreeNode(
-                            delegates_style=True,
-                            node_for=[MyDataItem],
-                            auto_open=True,
-                            label='text',
-                        ),
-                    ],
-                    hide_root=False,
-                    editable=False,
-                ),
-            ),
-            resizable=True,
-        )
-
-    MyClass().configure_traits()
